Add main files of the package
The package is the translate/translate.py file, which contains all the
functionalities of the package right now.

When the user ask for the translation of a word or an expression, the
code will perform the search on wordreference.com and return all the
translations found, organized in an array. The search is made with
requests, the html is parsed with BeautifulSoup, and the results are
formatted using terminaltables.